src/requirements.txt
arrow==1.3.0
asgiref==3.9.1
binaryornot==0.4.4
boto3==1.40.3
botocore==1.40.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.2
channels-redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
crispy-bootstrap5==2025.6
crispy-tailwind==1.0.3
cryptography==45.0.5
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-allauth-ui==1.8.1
django-crispy-forms==2.4
django-dotenv==1.4.2
django-environ==0.12.0
django-storages==1.14.6
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
groq==0.13.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
slippers==0.6.2
sqlparse==0.5.3
stripe==14.0.1
text-unidecode==1.3
typeguard==2.13.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.9.0
